+++ test/unit_tests/test_amz_cf_distribution.py
@@ -1,10 +1,12 @@
 #!/usr/bin/python3
 
 from amazonia.classes.cf_distribution_config import CFDistributionConfig, CFCacheBehaviorConfig, CFOriginsConfig
+from amazonia.classes.amz_cf_distribution import CFDistributionLeaf, CFDistributionUnit
+from network_setup import get_network_config
 from nose.tools import *
 
 
-def create_cf_distribution_config(aliases='wwwelb.ap-southeast-2.elb.amazonaws.com',
+def create_cf_distribution_config(aliases=['wwwelb.ap-southeast-2.elb.amazonaws.com'],
                                   comment='UnitTestCFDistConfig',
                                   default_root_object='index.html',
                                   enabled=True, price_class='PriceClass_All',
@@ -74,7 +76,7 @@ def create_custom_origin(domain_name='amazonia-elb-bucket.s3.amazonaws.com',
                          origin_protocol_policy='https-only',
                          http_port='80',
                          https_port='443',
-                         origin_ssl_protocols=('TLSv1', 'TLSv1.1', 'TLSv1.2')):
+                         origin_ssl_protocols=['TLSv1', 'TLSv1.1', 'TLSv1.2']):
     """
     Create a CustomOrigin object
     :param domain_name: The DNS name of the S3 bucket or HTTP server which this distribution will point to
@@ -106,16 +108,16 @@ def create_custom_origin(domain_name='amazonia-elb-bucket.s3.amazonaws.com',
 
 def create_cache_behavior(is_default=False,
                           path_pattern='/index.html',
-                          allowed_methods=('GET', 'POST'),
-                          cached_methods=('GET', 'POST'),
+                          allowed_methods=['GET', 'POST'],
+                          cached_methods=['GET', 'POST'],
                           target_origin_id='S3-bucket-id',
                           forward_cookies='all',
-                          forwarded_headers=('Accept', 'Set-Cookie'),
+                          forwarded_headers=['Accept', 'Set-Cookie'],
                           viewer_protocol_policy='allow-all',
                           min_ttl=0,
                           default_ttl=0,
                           max_ttl=0,
-                          trusted_signers=('self'),
+                          trusted_signers=['self'],
                           query_string=True):
     """
     :param path_pattern: The pattern to which this cache behavior applies
@@ -127,6 +129,7 @@ def create_cache_behavior(is_default=False,
     :param default_ttl: The default amount of time objects stay in the cache
     :param max_ttl: The maximum amount of time objects should stay in the cache
     :param forward_cookies: boolean to forward cookies to origin
+    :param forwarded_headers: list of headers to forward to origin
     :param trusted_signers: list of identifies that are trusted to sign cookies on behalf of this behavior
     :param query_string: indicates whether to forward query strings to the origin
     :return: Instance of CacheBehavior object
@@ -198,7 +201,7 @@ def test_custom_origin():
     origin_protocol_policy = 'https-only'
     http_port = '80'
     https_port = '443'
-    origin_ssl_protocols = ('TLSv1', 'TLSv1.1', 'TLSv1.2')
+    origin_ssl_protocols = ['TLSv1', 'TLSv1.1', 'TLSv1.2']
 
     helper_cf_origin = create_custom_origin(domain_name=domain_name,
                                             is_s3=is_s3,
@@ -222,16 +225,16 @@ def test_cf_cache_behavior():
     """
     is_default = False
     path_pattern = '/index.html'
-    allowed_methods = ('GET', 'POST')
-    cached_methods = ('GET', 'POST'),
+    allowed_methods = ['GET', 'POST']
+    cached_methods = ['GET', 'POST'],
     target_origin_id = 'S3-bucket-id'
     forward_cookies = 'all'
-    forwarded_headers = ('Accept', 'Set-Cookie')
+    forwarded_headers = ['Accept', 'Set-Cookie']
     viewer_protocol_policy = 'allow-all'
     min_ttl = 0
     default_ttl = 0
     max_ttl = 0
-    trusted_signers = ('self')
+    trusted_signers = ['self']
     query_string = True
 
     helper_cf_cache_behavior = create_cache_behavior(is_default=is_default,
@@ -267,7 +270,7 @@ def test_cf_distribution_config():
     Test to check DistributionConfig object inputs match the created outputs
     """
 
-    aliases = 'wwwelb.ap-southeast-2.elb.amazonaws.com'
+    aliases = ['wwwelb.ap-southeast-2.elb.amazonaws.com']
     comment = 'UnitTestCFDistConfig'
     default_root_object = 'index.html'
     enabled = True
@@ -277,10 +280,7 @@ def test_cf_distribution_config():
     minimum_protocol_version = 'TLSv1'
     ssl_support_method = 'sni-only'
 
-    origins = []
-    cache_behaviors = []
-
-    helper_cf_dist = create_cf_distribution_config(aliases=aliases,
+    cf_dist_config = create_cf_distribution_config(aliases=aliases,
                                                    comment=comment,
                                                    default_root_object=default_root_object,
                                                    enabled=enabled,
@@ -290,10 +290,51 @@ def test_cf_distribution_config():
                                                    minimum_protocol_version=minimum_protocol_version,
                                                    ssl_support_method=ssl_support_method)
 
-    assert_equal(aliases, helper_cf_dist.aliases)
-    assert_equal(comment, helper_cf_dist.comment)
-    assert_equal(default_root_object, helper_cf_dist.default_root_object)
-    assert_equal(enabled, helper_cf_dist.enabled)
-    assert_equal(price_class, helper_cf_dist.price_class)
-    assert_equal(acm_cert_arn, helper_cf_dist.acm_cert_arn)
-    assert_equal(error_page_path, helper_cf_dist.error_page_path)
+    assert_equal(aliases, cf_dist_config.aliases)
+    assert_equal(comment, cf_dist_config.comment)
+    assert_equal(default_root_object, cf_dist_config.default_root_object)
+    assert_equal(enabled, cf_dist_config.enabled)
+    assert_equal(price_class, cf_dist_config.price_class)
+    assert_equal(acm_cert_arn, cf_dist_config.acm_cert_arn)
+    assert_equal(error_page_path, cf_dist_config.error_page_path)
+
+
+def test_cf_distribution_unit():
+    """
+    Test CF distribution unit structure
+    """
+    network_config, template = get_network_config()
+    cf_dist_config = create_cf_distribution_config()
+    origins = [create_s3_origin(), create_s3_origin()]
+    default_behaviour = create_cache_behavior()
+    default_behaviour.is_default = True
+    cache_behaviors = [default_behaviour, create_cache_behavior()]
+    unit_title = 'testcf'
+    cf_dist_unit = CFDistributionUnit(unit_title=unit_title,
+                                      cf_distribution_config=cf_dist_config,
+                                      stack_config=network_config,
+                                      cf_origins_config=origins,
+                                      cf_cache_behavior_config=cache_behaviors,
+                                      template=template)
+    assert_equals(cf_dist_unit.title, unit_title)
+
+
+def test_cf_distribution_tree():
+    """
+    Test CF distribution leaf structure
+    """
+    network_config, template = get_network_config()
+    cf_dist_config = create_cf_distribution_config()
+    origins = [create_s3_origin(), create_s3_origin()]
+    default_behaviour = create_cache_behavior()
+    default_behaviour.is_default = True
+    cache_behaviors = [default_behaviour, create_cache_behavior()]
+    leaf_title = 'testcf'
+    tree_name = 'testtree'
+    cf_dist_unit = CFDistributionLeaf(leaf_title=leaf_title,
+                                      cf_distribution_config=cf_dist_config,
+                                      tree_name=tree_name,
+                                      cf_origins_config=origins,
+                                      cf_cache_behavior_config=cache_behaviors,
+                                      template=template)
+    assert_equals(cf_dist_unit.title, leaf_title)