+++ test/unit_tests/test_amz_zd_autoscaling.py
@@ -1,4 +1,4 @@
-from amazonia.classes.amz_zd_autoscaling import ZdAutoscalingUnit
+from amazonia.classes.amz_zd_autoscaling import ZdAutoscalingUnit, ZdAutoscalingLeaf
 from amazonia.classes.asg_config import AsgConfig
 from amazonia.classes.block_devices_config import BlockDevicesConfig
 from amazonia.classes.elb_config import ElbConfig, ElbListenersConfig
@@ -6,12 +6,21 @@ from network_setup import get_network_config
 from nose.tools import *
 from troposphere import Ref
 
-template = elb_config = network_config = common_asg_config = block_devices_config = None
+template = elb_config = network_config = common_asg_config = block_devices_config = tree_name = availability_zones = \
+    cd_service_role_arn = public_cidr = public_hosted_zone_name = keypair = None
 
 
 def setup_resources():
     """ Setup global variables between tests"""
-    global template, elb_config, network_config, common_asg_config, block_devices_config
+    global template, elb_config, network_config, common_asg_config, block_devices_config, tree_name, \
+        availability_zones, cd_service_role_arn, public_cidr, public_hosted_zone_name, keypair
+    tree_name = 'testtree'
+    network_config, template = get_network_config()
+    availability_zones = ['ap-southeast-2a', 'ap-southeast-2b', 'ap-southeast-2c']
+    cd_service_role_arn = 'arn:aws:iam::123456789:role/CodeDeployServiceRole'
+    public_cidr = {'name': 'PublicIp', 'cidr': '0.0.0.0/0'}
+    public_hosted_zone_name = 'your.domain.'
+    keypair = 'INSERT_YOUR_KEYPAIR_HERE'
     block_devices_config = [BlockDevicesConfig(
         device_name='/dev/xvda',
         ebs_volume_size='15',
@@ -68,11 +77,10 @@ runcmd:
 @with_setup(setup_resources)
 def test_autoscaling_unit():
     """Test zdtd autoscaling unit structure"""
-    global common_asg_config
     title = 'app'
     blue_asg_config = common_asg_config
     green_asg_config = common_asg_config
-    unit = create_zdtd_autoscaling_unit(unit_title=title,
+    unit = create_zdtd_autoscaling_unit(title=title,
                                         blue_asg_config=blue_asg_config,
                                         green_asg_config=green_asg_config)
     assert_equals(unit.green_asg.trop_asg.title, 'green' + title + 'Asg')
@@ -95,15 +103,14 @@ def test_autoscaling_unit():
 @with_setup(setup_resources)
 def test_unit_association():
     """Test zdtd autoscaling unit flow"""
-    global common_asg_config
     blue_asg_config = common_asg_config
 
     green_asg_config = common_asg_config
-    unit1 = create_zdtd_autoscaling_unit(unit_title='app1',
+    unit1 = create_zdtd_autoscaling_unit(title='app1',
                                          blue_asg_config=blue_asg_config,
                                          green_asg_config=green_asg_config,
                                          dependencies=['app2:80'])
-    create_zdtd_autoscaling_unit(unit_title='app2',
+    create_zdtd_autoscaling_unit(title='app2',
                                  blue_asg_config=blue_asg_config,
                                  green_asg_config=green_asg_config)
 
@@ -117,16 +124,62 @@ def test_unit_association():
     assert_equals(len(unit1.prod_elb.egress), 2)
 
 
-def create_zdtd_autoscaling_unit(unit_title, blue_asg_config, green_asg_config, dependencies=None):
+@with_setup(setup_resources)
+def test_autoscaling_leaf():
+    """Test zdtd autoscaling leaf structure"""
+    title = 'app'
+    blue_asg_config = common_asg_config
+    green_asg_config = common_asg_config
+    leaf = create_zdtd_autoscaling_leaf(title=title,
+                                        blue_asg_config=blue_asg_config,
+                                        green_asg_config=green_asg_config)
+    assert_equals(leaf.green_asg.trop_asg.title, 'green' + title + 'Asg')
+    assert_equals(leaf.blue_asg.trop_asg.title, 'blue' + title + 'Asg')
+    assert_equals(leaf.prod_elb.trop_elb.title, title)
+    assert_equals(leaf.pre_elb.trop_elb.title, 'pre' + title)
+    [assert_is(type(lbn), Ref) for lbn in leaf.green_asg.trop_asg.LoadBalancerNames]
+    [assert_is(type(lbn), Ref) for lbn in leaf.blue_asg.trop_asg.LoadBalancerNames]
+
+    assert_equals(len(leaf.blue_asg.egress), 1)
+    assert_equals(len(leaf.green_asg.egress), 1)
+    assert_equals(len(leaf.blue_asg.ingress), 3)
+    assert_equals(len(leaf.green_asg.ingress), 3)
+    assert_equals(len(leaf.prod_elb.ingress), 1)
+    assert_equals(len(leaf.pre_elb.ingress), 1)
+    assert_equals(len(leaf.prod_elb.egress), 2)
+    assert_equals(len(leaf.pre_elb.egress), 2)
+
+
+@with_setup(setup_resources)
+def test_leaf_association():
+    """Test zdtd autoscaling leaf flow"""
+    blue_asg_config = common_asg_config
+
+    green_asg_config = common_asg_config
+    leaf = create_zdtd_autoscaling_unit(title='app1',
+                                        blue_asg_config=blue_asg_config,
+                                        green_asg_config=green_asg_config,
+                                        dependencies=['app2:80'])
+
+    assert_equals(len(leaf.blue_asg.egress), 2)
+    assert_equals(len(leaf.green_asg.egress), 2)
+    assert_equals(len(leaf.blue_asg.ingress), 3)
+    assert_equals(len(leaf.green_asg.ingress), 3)
+    assert_equals(len(leaf.pre_elb.ingress), 1)
+    assert_equals(len(leaf.prod_elb.ingress), 1)
+    assert_equals(len(leaf.pre_elb.egress), 2)
+    assert_equals(len(leaf.prod_elb.egress), 2)
+
+
+def create_zdtd_autoscaling_unit(title, blue_asg_config, green_asg_config, dependencies=None):
     """Helper function to create unit
-    :param unit_title: title of unit
+    :param title: title of unit
     :param blue_asg_config: blue specific asg config
     :param green_asg_config: green specific asg config
     :return new zdtd_autoscaling unit
     """
-    global template, elb_config, network_config
-    unit = ZdAutoscalingUnit(
-        unit_title=unit_title,
+    return ZdAutoscalingUnit(
+        unit_title=title,
         template=template,
         dependencies=dependencies,
         blue_asg_config=blue_asg_config,
@@ -134,4 +187,12 @@ def create_zdtd_autoscaling_unit(unit_title, blue_asg_config, green_asg_config, 
         elb_config=elb_config,
         stack_config=network_config
     )
-    return unit
+
+
+def create_zdtd_autoscaling_leaf(title, blue_asg_config, green_asg_config, dependencies=None):
+    global template, elb_config
+    return ZdAutoscalingLeaf(tree_name=tree_name, leaf_title=title, blue_asg_config=blue_asg_config,
+                             green_asg_config=green_asg_config, elb_config=elb_config,
+                             availability_zones=availability_zones, cd_service_role_arn=cd_service_role_arn,
+                             keypair=keypair, template=template, public_cidr=public_cidr,
+                             dependencies=dependencies, public_hosted_zone_name=public_hosted_zone_name)